// frontend_admin/types/supabase.ts
export type Database = {
  public: {
    Tables: {
// ...
      files: {
        Row: {
          created_at: string | null
          description: string | null
          extension: string | null
          extracted_data: string | null
          file_id: number
          file_name: string | null
          link: string | null
          project_id: number | null
          title: string | null
          type: string | null
          updated_at: string | null
          user_id: string | null
        }
        Insert: {
          created_at?: string | null
          description?: string | null
          extension?: string | null
          extracted_data?: string | null
          file_id?: never
          file_name?: string | null
          link?: string | null
          project_id?: number | null
          title?: string | null
          type?: string | null
          updated_at?: string | null
          user_id?: string | null
        }
        Update: {
          created_at?: string | null
          description?: string | null
          extension?: string | null
          extracted_data?: string | null
          file_id?: never
          file_name?: string | null
          link?: string | null
          project_id?: number | null
          title?: string | null
          type?: string | null
          updated_at?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "files_project_id_fkey"
            columns: ["project_id"]
            isOneToOne: false
            referencedRelation: "projects"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "files_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}